Update dependency tracking
The README now provides a command to inspect all the changes made to
hostap. This assures that this diff is always up to date, meaning in
subsequent commits we can directly modify the hostap code without always
needing to update the diff.

The wpaspy.py module is now a symbolic link to the file in the hostap
project to avoid duplication and to keep it easier to keep this file up
to date.

The repository also also been updated to track the latest libwifi
version which includes new functions to implement the 4-way handshake in
python.