hostapd: add GetIp action
The GetIp action causes the AP to start acting as a DHCP server.
Once the client requested an AP the action completes. This is based
on the code of FragAttacks, with the difference that the DHCP server
is only initialized once the GetIp action is started.

One part that is not yet implemented is manually assigning IPs, in
which cause the GetIp action would do nothing.